requirements.txt
Pillow==8.1.0
asgiref==3.2.10
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1
django-ckeditor==6.0.0
django-etc==1.3.0
django-import-export==2.5.0
django-js-asset==1.2.2
django-user-agents==0.4.0
et-xmlfile==1.0.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
psycopg2==2.8.6
python-memcached==1.59
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
ua-parser==0.10.0
user-agents==2.2.0
xlrd==2.0.1
xlwt==1.3.0
